refactor: split CrtDisplay into partial classes by responsibility
CrtDisplay.cs was a single 819-line file. Split into three focused
partial class files:
- CrtDisplay.cs: core logic (overlay layer, state, lifecycle, timers, noise)
- CrtDisplay.Properties.cs: StyledProperty declarations and CLR wrappers
- CrtDisplay.Rendering.cs: Draw* effect methods and brush cache helpers

No behaviour changes.
